Moved database related settings to databaseManager
Since there was no reason to have a separate process that only reads
the configuration file for the settings related to the database
and sends them to dbus to the databaseManager I moved that part
into the databaseManager.

SimpleCrypt class was also moved to the mailManager since there will
used directly.