Added mailManager dbus interface
In order to be able to send mail on alarm events the dbus interface
for the mailManager needed to be implemented.
The sensorInstantiator will only connect and login to smtp server
once, then each type of sensor(wired and wireless) will be able
to send a custom mail to the configured list of recipients from
the mailManager settings file.