Added first time initialization
When the applications start, will check if the configuration files
are in place, if not will consider first time initialization and
will generate and save some default configurations for first use.
After the .ini file is generated the application will exit giving
the user the change to customize the configurations.
At the next application start, the configuration will be loaded
and the application will be initialized normally.