Added utils library that provides some nice features
This commit adds implementation for the signal handlers. Each application
can now link with the piHomeUtils library and select what signals to be
handled.
In the special case of SIGSEGV there is a custom handler that also tries
to demangle the symbols in order to print the real function name instead
of some hex values and offsets.
The logging was also added in this library, each application will have
all messages saved to a log file with the application name in
~/.piHome/logging directory.