Added systemd service files and moved to system dbus
Added for each application a systemd service file in order to have
the application started at boot.
Moved all applications from session bus to system bus for achieving
a higher level of security. For each application that is posting a
DBus service, a configuration file is needed. All applications are
started by the root user.
For the systemd service files and for the dbus configuration files,
install rules were added in each project file.