Added support for systemd watchdog in utils library
Common functionality for implementing the systemd watchdog was
added in the utils library. Each component should link and use
this as reference.
The DatabaseManager, MailManager and newly implemented
SensorManager implemented the interface provided by the utils
library and the service files contain the specific information
in order to enable the systemd watchdog.