Added actuator/sensor table in database
In order to have a more rapid way to get different sensors/actuators
related information in a compact form.
When the sensors/actuators are instantiated, an entry in the IO table
is inserted. The IO table will contain only unique information, so it's
not possible to insert the same information multiple times.
Implemented the getAllZones method that is exposed to DBus. The
databaseManagerInterface and the mailManagerInterface were moved to the
piHomeCommon because will be used by other interfaces also.