Added userManager module
The userManager module will add the login functionality for the
application.
The user list will be read from the users.ini configuration
file. Each user has defined an username, a pin and a password.
A list of MAC addresses is also defined for each user.
The simpleCrypt was moved here because is used for password and
pin encryption/decryption.