Changed UIManager based on AlarmManager and UserManager
UIManager was updated to support the functionality provided by the
newly introduced AlarmManager and UserManager.
UserList is read from the UserManager, the passwords and pins are
validated via dbus.
If login timeout expires, the UI will not allow changing screens or
get access to any kind of information.
The three alarm states are updated based on the UI settings. The UI
is initialized with the last state of the alarm, to avoid changing
the state in case of failure.