Added webManager for controling the system remotely
WebManager is based on QtWebApp library from Stephan Frings.
The implementation is based on userManager for providing the
login support.
The alarmManager provides support for arming/disarming the
system or switching between the alarm modes.
The sensorManager provides the list of events and the functionality
for locking/unlocking the front door.
Everything is controlled remotely after login to an web page, that
is mobile friendly and provides basic interraction with the system.